+++ lib/project_types/rails/cli.rb
@@ -5,17 +5,10 @@ module Rails
     register_messages(Rails::Messages::MESSAGES)
   end
 
-  # define/autoload project specific Commands
-  class Command < ShopifyCLI::ProjectCommands
-    subcommand :Connect, "connect", Project.project_filepath("commands/connect")
-    subcommand :Create, "create", Project.project_filepath("commands/create")
-    subcommand :Deploy, "deploy", Project.project_filepath("commands/deploy")
-    subcommand :Generate, "generate", Project.project_filepath("commands/generate")
-    subcommand :Open, "open", Project.project_filepath("commands/open")
-    subcommand :Serve, "serve", Project.project_filepath("commands/serve")
-    subcommand :Tunnel, "tunnel", Project.project_filepath("commands/tunnel")
+  class Command
+    autoload :Create, Project.project_filepath("commands/create")
+    autoload :Generate, Project.project_filepath("commands/generate")
   end
-  ShopifyCLI::Commands.register("Rails::Command", "rails")
 
   # define/autoload project specific Tasks
   module Tasks