+++ lib/project_types/script/commands/create.rb
@@ -2,7 +2,7 @@
 
 module Script
   class Command
-    class Create < ShopifyCLI::SubCommand
+    class Create < ShopifyCLI::Command::SubCommand
       unless ShopifyCLI::Environment.acceptance_test?
         prerequisite_task :ensure_authenticated
       end