+++ lib/shopify_cli/command/app_sub_command.rb
@@ -1,6 +1,12 @@
 module ShopifyCLI
   class Command
     class AppSubCommand < SubCommand
+      def detect_app(directory: Dir.pwd)
+        AppTypeDetector.detect(project_directory: directory)
+      rescue ShopifyCLI::AppTypeDetector::TypeNotFoundError
+        raise ShopifyCLI::Abort, @ctx.message("core.app.type_not_found", directory)
+      end
+
       class << self
         def call_help(*)
           output = help